+++ scripts/publishCI.ts
@@ -1,7 +1,6 @@
 import { publish } from '@vitejs/release-scripts'
 
 publish({
-  provenance: true,
   getPkgDir(pkg) {
     if (pkg === 'plugin-react-swc') {
       return `packages/${pkg}/dist`